@@ -1,12 +1,12 @@
 allprojects {
-    group 'com.github.pozo'
-    version '1.4.0.0'
+    group 'io.github.dhanfinix'
+    version '1.0.0-alpha'
 }
 
 buildscript {
     ext {
-        kotlin_version = '1.3.72'
-        kotlinxMetadataVersion = '0.0.5'
+        kotlin_version = '1.9.22'
+        kotlinxMetadataVersion = '0.9.0'
         kotlinBuilderAnnotationVersion = '1.0-SNAPSHOT'
         autoServiceVersion = '1.0-rc7'
         mapstructVersion = '1.4.0.Beta3'
