@@ -1,56 +1,63 @@
-group 'com.github.pozo'
-version '1.0-SNAPSHOT'
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 buildscript {
-    ext {
-        kotlin_version = '1.3.72'
-        mapstructVersion = '1.3.1.Final'
-        mapstructKotlinVersion = '1.3.1.2'
-    }
+    ext.kotlin_version = '1.9.22' // Updated Kotlin version
+    ext.mapstructVersion = '1.3.1.Final' // Latest MapStruct version
+    ext.mapstructKotlinVersion = '1.3.1.2' // Updated MapStruct Kotlin version
+    ext.kotlin_mapstruct = '1.0.0-alpha'
 
     repositories {
         mavenCentral()
-        maven {
-            url('https://plugins.gradle.org/m2/')
-        }
+        gradlePluginPortal()
     }
 
     dependencies {
-        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version")
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
 
-apply plugin: 'kotlin'
-apply plugin: 'kotlin-kapt'
+plugins {
+    id 'org.jetbrains.kotlin.jvm' version "$kotlin_version"
+    id 'org.jetbrains.kotlin.kapt' version "$kotlin_version"
+}
 
-sourceCompatibility = JavaVersion.VERSION_1_8
+group = 'com.github.pozo'
+version = '1.0-SNAPSHOT'
+
+java {
+    toolchain {
+        languageVersion.set(JavaLanguageVersion.of(17)) // Updated to match your Java version
+    }
+}
 
 repositories {
     mavenLocal()
     mavenCentral()
     maven { url "https://oss.sonatype.org/content/repositories/snapshots/" }
-    maven { url "https://kotlin.bintray.com/kotlinx" }
 }
 
 dependencies {
     implementation project(":domain")
 
-    api("com.github.pozo:mapstruct-kotlin:${mapstructKotlinVersion}")
-    api("org.mapstruct:mapstruct:${mapstructVersion}")
+    api "io.github.dhanfinix:mapstruct-kotlin:$kotlin_mapstruct"
+    api "org.mapstruct:mapstruct:$mapstructVersion"
 
-    kapt("org.mapstruct:mapstruct-processor:${mapstructVersion}")
-    kapt("com.github.pozo:mapstruct-kotlin-processor:${mapstructKotlinVersion}")
+    kapt "org.mapstruct:mapstruct-processor:$mapstructVersion"
+    kapt "io.github.dhanfinix:mapstruct-kotlin-processor:$kotlin_mapstruct"
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 
     testImplementation "org.jetbrains.kotlin:kotlin-test:$kotlin_version"
-    testImplementation 'junit:junit:4.12'
+    testImplementation 'junit:junit:4.13.2' // Updated JUnit version
 }
 
-compileKotlin {
-    kotlinOptions.jvmTarget = JavaVersion.VERSION_1_8
-}
-compileTestKotlin {
-    kotlinOptions.jvmTarget = JavaVersion.VERSION_1_8
+tasks.withType(KotlinCompile).configureEach {
+    compilerOptions {
+        jvmTarget.set(JvmTarget.JVM_17)
+    }
 }
 
+tasks.withType(JavaCompile).configureEach {
+    options.release.set(17)
+}
