@@ -1,98 +1,150 @@
-apply plugin: 'java-library'
-apply plugin: 'kotlin'
-apply plugin: 'kotlin-kapt'
-apply plugin: 'maven-publish'
-apply plugin: 'signing'
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+import com.vanniktech.maven.publish.AndroidSingleVariantLibrary
+import com.vanniktech.maven.publish.SonatypeHost
 
-sourceCompatibility = JavaVersion.VERSION_1_8
+plugins {
+    id 'java-library'
+    id 'kotlin'
+    id 'kotlin-kapt'
+    id 'maven-publish'
+    id 'signing'
+    id "com.vanniktech.maven.publish" version "0.32.0"
+}
 
 repositories {
     mavenCentral()
-    maven { url "https://kotlin.bintray.com/kotlinx" }
 }
 
 dependencies {
-    implementation("org.mapstruct:mapstruct:${mapstructVersion}")
+    implementation "org.mapstruct:mapstruct:${mapstructVersion}"
     implementation "org.mapstruct:mapstruct-processor:${mapstructVersion}"
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib"
     implementation "org.jetbrains.kotlinx:kotlinx-metadata-jvm:${kotlinxMetadataVersion}"
     implementation "com.google.auto.service:auto-service:${autoServiceVersion}"
     implementation "com.squareup:javapoet:${javaPoetVersion}"
 
-    implementation(project(':mapstruct-kotlin-builder'))
+    implementation project(':mapstruct-kotlin-builder')
 
     kapt "com.google.auto.service:auto-service:${autoServiceVersion}"
 
     testImplementation 'org.mockito:mockito-core:2.28.2'
-    testImplementation 'junit:junit:4.12'
+    testImplementation 'junit:junit:4.13.2' // Updated JUnit version
 }
 
-task sourcesJar(type: Jar, dependsOn: classes) {
-    classifier = 'sources'
+java {
+    toolchain {
+        languageVersion.set(JavaLanguageVersion.of(17)) // Updated to match your Java version
+    }
+}
+
+tasks.register("sourcesJar", Jar) {
+    dependsOn classes
+    archiveClassifier.set("sources")
     from sourceSets.main.allSource
 }
 
-task javadocJar(type: Jar, dependsOn: javadoc) {
-    classifier = 'javadoc'
+tasks.register("javadocJar", Jar) {
+    dependsOn javadoc
+    archiveClassifier.set("javadoc")
     from javadoc.destinationDir
 }
 
-artifacts {
-    archives sourcesJar
-    archives javadocJar
-}
+mavenPublishing {
+    publishToMavenCentral(SonatypeHost.CENTRAL_PORTAL)
+    signAllPublications()
 
-publishing {
-    publications {
-        maven(MavenPublication) {
-            from components.java
+    coordinates(group, "mapstruct-kotlin-processor", version)
 
-            artifact sourcesJar
-            artifact javadocJar
-            pom {
-                name = 'mapstruct-kotlin-processor'
-                description = 'KotlinBuilder annotation processor'
-                url = 'https://github.com/Pozo/mapstruct-kotlin'
-                licenses {
-                    license {
-                        name = 'MIT License'
-                        url = 'https://opensource.org/licenses/MIT'
-                    }
-                }
-                developers {
-                    developer {
-                        id = 'pozo'
-                        name = 'Zoltan Polgar'
-                        email = 'pozo@gmx.com'
-                    }
-                }
-                scm {
-                    url = 'https://github.com/Pozo/mapstruct-kotlin/tree/master'
-                    connection = 'scm:git:git://github.com/Pozo/mapstruct-kotlin.git'
-                    developerConnection = 'scm:git:ssh://github.com:Pozo/mapstruct-kotlin.git'
-                }
+    pom {
+        name = "mapstruct-kotlin-processor"
+        description = "KotlinBuilder annotation processor plugin for MapStruct"
+        url = "https://github.com/Dhanfinix/mapstruct-kotlin"
+
+        licenses {
+            license {
+                name = "MIT License"
+                url = "https://opensource.org/licenses/MIT"
             }
         }
-    }
-    repositories {
-        maven {
-            url "https://oss.sonatype.org/service/local/staging/deploy/maven2/"
-            credentials {
-                username "${nexusUsername}"
-                password "${nexusPassword}"
+        developers {
+            developer {
+                id = "dhanfinix"
+                name = "Ramdhan"
+                email = "muhammadramdhan541@gmail.com"
             }
+        }
+        scm {
+            url = "https://github.com/Dhanfinix/mapstruct-kotlin"
+            connection = "scm:git:git://github.com/Dhanfinix/mapstruct-kotlin.git"
+            developerConnection = "scm:git:ssh://github.com:Dhanfinix/mapstruct-kotlin.git"
         }
     }
 }
+//publishing {
+//    publications {
+//        maven(MavenPublication) {
+//            artifactId = "mapstruct-kotlin-processor"
+//
+//            from(components["java"])
+//            artifact(tasks.named("sourcesJar").get())
+//            artifact(tasks.named("javadocJar").get())
+//
+//            pom {
+//                name = "mapstruct-kotlin-processor"
+//                description = "Fork of KotlinBuilder annotation processor for MapStruct"
+//                url = "https://github.com/Dhanfinix/mapstruct-kotlin"
+//
+//                licenses {
+//                    license {
+//                        name = "MIT License"
+//                        url = "https://opensource.org/licenses/MIT"
+//                    }
+//                }
+//                developers {
+//                    developer {
+//                        id = "dhanfinix"
+//                        name = "Ramdhan"
+//                        email = "muhammadramdhan541@gmail.com"
+//                    }
+//                }
+//                scm {
+//                    url = "https://github.com/Dhanfinix/mapstruct-kotlin"
+//                    connection = "scm:git:git://github.com/Dhanfinix/mapstruct-kotlin.git"
+//                    developerConnection = "scm:git:ssh://github.com:Dhanfinix/mapstruct-kotlin.git"
+//                }
+//            }
+//        }
+//    }
+//    repositories {
+//        maven {
+//            url "https://oss.sonatype.org/service/local/staging/deploy/maven2/"
+//            credentials {
+//                username = findProperty("credential_name")
+//                password = findProperty("credential_pass")
+//            }
+//        }
+//    }
+//}
 
+//signing {
+////    useGpgCmd()
+//    useInMemoryPgpKeys(
+//        findProperty("signing.keyId") as String,
+//        findProperty("signing.secretKey") as String,
+//        findProperty("signing.password") as String
+//    )
+//    sign publishing.publications.maven
+//}
 signing {
-    useGpgCmd()
-    sign publishing.publications.maven
+    useInMemoryPgpKeys(
+            findProperty("signing.keyId") as String,
+            findProperty("signing.secretKey") as String,
+            findProperty("signing.password") as String
+    )
+    sign(publishing.publications)
 }
 
-compileKotlin {
-    kotlinOptions.jvmTarget = JavaVersion.VERSION_1_8
-}
-compileTestKotlin {
-    kotlinOptions.jvmTarget = JavaVersion.VERSION_1_8
+
+tasks.withType(KotlinCompile).configureEach {
+    kotlinOptions.jvmTarget = "17"
 }
